Refactor: optimize physics engine, clean up codebase, and improve type safety
- Pre-allocate arrays in PhysicsEngineRk4 to eliminate ~3300 GC allocations/frame
- Inline RK4 K1-K4 calculations for clarity and to support array reuse
- Optimize CalcAccelerations: use G*rij/r³ instead of G/r²*normalize(rij), reducing sqrt calls from 3 to 1 per pair
- Fix precision loss: remove float casts in energy calculations, use full double precision
- Change Id/ParentId from float to int across the entire project (Astro, AstroConfig, AstroData, Camera, JSON data)
- Satellite IDs remapped: 3.1→31, 4.1→41, 5.1→51, etc.
- Change _n (RK4 subdivisions) from double to int for correct semantics
- Pre-calculate subStep (_timeStep/_n) instead of recalculating every frame
- Extract UpdateDisplayValues from Draw to separate state logic from rendering
- Remove unused parameters from DrawInfoText
- Fix double Math.Sin calculation in DrawCross
- Remove redundant null check in DrawRings
- Replace magic numbers in EnergyMenu with lineHeight loop
- Implement grid-based layout in HelpMenu replacing manual coordinate math
- Move screenBounds outside foreach loop in DrawBodies
- Cache Sun reference in RenderSystem to avoid LINQ search every frame
- Use TryGetValue instead of ContainsKey+indexer in InputSystems
- Fix Vector2D.Normalize to calculate Length() once instead of twice
- Remove unused Vector2D methods (Add, Subtract, Multiply, Divide, Negate, instance overloads)
- Remove unused operator overloads (float, decimal, vector-vector)
- Clean up commented-out logger code
- Remove unnecessary discard in Trail.Dequeue()
- Remove static modifier from _keyName in Simulation

+++ NBodiesSim/Source/Core/Camera.cs
@@ -24,7 +24,7 @@ internal class Camera
     // as well as changes in the chosen body.
     public double TargetDistanceScale { get; set; } = CameraConstants.DefaultDistanceScale;
     public double TargetRadiusScale { get; set; } = CameraConstants.DefaultRadiusScale;
-    public float TargetId { get; set; } // ID of the body to follow
+    public int TargetId { get; set; } // ID of the body to follow
 
     // Interpolation configuration
     private double LerpSpeed { get; set; } = CameraConstants.InitialLerpSpeed;
